+++ src/mathematics/Quaternion.h
@@ -30,6 +30,7 @@
 #include <cmath>
 #include "Vector3.h"
 #include "Matrix3x3.h"
+#include "../decimal.h"
 
 // ReactPhysics3D namespace
 namespace reactphysics3d {
@@ -42,86 +43,86 @@ namespace reactphysics3d {
 */
 class Quaternion {
     private :
-        double x;                       // Component x of the quaternion
-        double y;                       // Component y of the quaternion
-        double z;                       // Component z of the quaternion
-        double w;                       // Component w of the quaternion
+        decimal x;      // Component x of the quaternion
+        decimal y;      // Component y of the quaternion
+        decimal z;      // Component z of the quaternion
+        decimal w;      // Component w of the quaternion
 
     public :
-        Quaternion();                                                       // Constructor
-        Quaternion(double x, double y, double z, double w);                 // Constructor with arguments
-        Quaternion(double w, const Vector3& v);                             // Constructor with the component w and the vector v=(x y z)
-        Quaternion(const Quaternion& quaternion);                           // Copy-constructor
-        Quaternion(const Matrix3x3& matrix);                                // Create a unit quaternion from a rotation matrix
-        ~Quaternion();                                                      // Destructor
-        double getX() const;                                                // Return the component x of the quaternion
-        double getY() const;                                                // Return the component y of the quaternion
-        double getZ() const;                                                // Return the component z of the quaternion
-        double getW() const;                                                // Return the component w of the quaternion
-        void setX(double x);                                                // Set the value x
-        void setY(double y);                                                // Set the value y
-        void setZ(double z);                                                // Set the value z
-        void setW(double w);                                                // Set the value w
-        Vector3 vectorV() const;                                            // Return the vector v=(x y z) of the quaternion
-        double length() const;                                              // Return the length of the quaternion
-        Quaternion getUnit() const;                                         // Return the unit quaternion
-        Quaternion getConjugate() const;                                    // Return the conjugate quaternion
-        Quaternion getInverse() const;                                      // Return the inverse of the quaternion
-        Matrix3x3 getMatrix() const;                                        // Return the orientation matrix corresponding to this quaternion
-        static Quaternion identity();                                       // Return the identity quaternion
-        double dot(const Quaternion& quaternion) const;                     // Dot product between two quaternions
-        void getRotationAngleAxis(double& angle, Vector3& axis) const;      // Compute the rotation angle (in radians) and the axis
+        Quaternion();                                                   // Constructor
+        Quaternion(decimal x, decimal y, decimal z, decimal w);         // Constructor with arguments
+        Quaternion(decimal w, const Vector3& v);                        // Constructor with the component w and the vector v=(x y z)
+        Quaternion(const Quaternion& quaternion);                       // Copy-constructor
+        Quaternion(const Matrix3x3& matrix);                            // Create a unit quaternion from a rotation matrix
+        ~Quaternion();                                                  // Destructor
+        decimal getX() const;                                           // Return the component x of the quaternion
+        decimal getY() const;                                           // Return the component y of the quaternion
+        decimal getZ() const;                                           // Return the component z of the quaternion
+        decimal getW() const;                                           // Return the component w of the quaternion
+        void setX(decimal x);                                           // Set the value x
+        void setY(decimal y);                                           // Set the value y
+        void setZ(decimal z);                                           // Set the value z
+        void setW(decimal w);                                           // Set the value w
+        Vector3 vectorV() const;                                        // Return the vector v=(x y z) of the quaternion
+        decimal length() const;                                         // Return the length of the quaternion
+        Quaternion getUnit() const;                                     // Return the unit quaternion
+        Quaternion getConjugate() const;                                // Return the conjugate quaternion
+        Quaternion getInverse() const;                                  // Return the inverse of the quaternion
+        Matrix3x3 getMatrix() const;                                    // Return the orientation matrix corresponding to this quaternion
+        static Quaternion identity();                                   // Return the identity quaternion
+        decimal dot(const Quaternion& quaternion) const;                // Dot product between two quaternions
+        void getRotationAngleAxis(decimal& angle, Vector3& axis) const; // Compute the rotation angle (in radians) and the axis
         static Quaternion slerp(const Quaternion& quaternion1,
-                                const Quaternion& quaternion2, double t);   // Compute the spherical linear interpolation between two quaternions
+                                const Quaternion& quaternion2, decimal t);   // Compute the spherical linear interpolation between two quaternions
 
         // --- Overloaded operators --- //
-        Quaternion operator+(const Quaternion& quaternion) const;             // Overloaded operator for the addition
-        Quaternion operator-(const Quaternion& quaternion) const;             // Overloaded operator for the substraction
-        Quaternion operator*(double nb) const;                                // Overloaded operator for the multiplication with a constant
-        Quaternion operator*(const Quaternion& quaternion) const;             // Overloaded operator for the multiplication
-        Quaternion& operator=(const Quaternion& quaternion);                  // Overloaded operator for assignment
-        bool operator==(const Quaternion& quaternion) const;                  // Overloaded operator for equality condition
+        Quaternion operator+(const Quaternion& quaternion) const;       // Overloaded operator for the addition
+        Quaternion operator-(const Quaternion& quaternion) const;       // Overloaded operator for the substraction
+        Quaternion operator*(decimal nb) const;                         // Overloaded operator for the multiplication with a constant
+        Quaternion operator*(const Quaternion& quaternion) const;       // Overloaded operator for the multiplication
+        Quaternion& operator=(const Quaternion& quaternion);            // Overloaded operator for assignment
+        bool operator==(const Quaternion& quaternion) const;            // Overloaded operator for equality condition
 };
 
 // --- Inline functions --- //
 
 // Get the value x (inline)
-inline double Quaternion::getX() const {
+inline decimal Quaternion::getX() const {
     return x;
 }
 
 // Get the value y (inline)
-inline double Quaternion::getY() const {
+inline decimal Quaternion::getY() const {
     return y;
 }
 
 // Get the value z (inline)
-inline double Quaternion::getZ() const {
+inline decimal Quaternion::getZ() const {
     return z;
 }
 
 // Get the value w (inline)
-inline double Quaternion::getW() const {
+inline decimal Quaternion::getW() const {
     return w;
 }
 
 // Set the value x (inline)
-inline void Quaternion::setX(double x) {
+inline void Quaternion::setX(decimal x) {
     this->x = x;
 }
 
 // Set the value y (inline)
-inline void Quaternion::setY(double y) {
+inline void Quaternion::setY(decimal y) {
     this->y = y;
 }
 
 // Set the value z (inline)
-inline void Quaternion::setZ(double z) {
+inline void Quaternion::setZ(decimal z) {
     this->z = z;
 }
 
 // Set the value w (inline)
-inline void Quaternion::setW(double w) {
+inline void Quaternion::setW(decimal w) {
     this->w = w;
 }
 
@@ -132,13 +133,13 @@ inline Vector3 Quaternion::vectorV() const {
 }
 
 // Return the length of the quaternion (inline)
-inline double Quaternion::length() const {
+inline decimal Quaternion::length() const {
     return sqrt(x*x + y*y + z*z + w*w);
 }
 
 // Return the unit quaternion
 inline Quaternion Quaternion::getUnit() const {
-    double lengthQuaternion = length();
+    decimal lengthQuaternion = length();
 
     // Check if the length is not equal to zero
     assert (lengthQuaternion != 0.0);
@@ -159,7 +160,7 @@ inline Quaternion Quaternion::getConjugate() const {
 
 // Return the inverse of the quaternion (inline)
 inline Quaternion Quaternion::getInverse() const {
-    double lengthQuaternion = length();
+    decimal lengthQuaternion = length();
     lengthQuaternion = lengthQuaternion * lengthQuaternion;
 
     assert (lengthQuaternion != 0.0);
@@ -169,7 +170,7 @@ inline Quaternion Quaternion::getInverse() const {
 }
 
 // Scalar product between two quaternions
-inline double Quaternion::dot(const Quaternion& quaternion) const {
+inline decimal Quaternion::dot(const Quaternion& quaternion) const {
     return (x*quaternion.x + y*quaternion.y + z*quaternion.z + w*quaternion.w);
 }
 
@@ -186,7 +187,7 @@ inline Quaternion Quaternion::operator-(const Quaternion& quaternion) const {
 }
 
 // Overloaded operator for the multiplication with a constant
-inline Quaternion Quaternion::operator*(double nb) const {
+inline Quaternion Quaternion::operator*(decimal nb) const {
     // Return the result
     return Quaternion(nb*x, nb*y, nb*z, nb*w);
 }
